# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Backend API for Report It, a campus incident reporting platform. Students register with their college email, verify it over an email OTP, and file reports such as lost and found, harassment, ragging or canteen overcharging. Admins can moderate users and posts.",
        "title": "Report It API",
        "contact": {
            "name": "Report It API Support"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/authorized/admin/remove-post/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes the report post identified by the path id. The caller role is taken from the access token and the owning user id is taken from the request body. Registered both as a user route and as an admin route.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts"
                ],
                "summary": "Delete a post",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Post id (Mongo ObjectID hex)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Owning user id",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/post.DeletePostRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message and the deleted post.Post",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "role missing from the context, invalid JSON body, or deletion failed",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/post": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a paginated list of every report post. The page query parameter is read as Page (capital P) and defaults to 1; limit defaults to 10 and is clamped to the range 1-100.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts"
                ],
                "summary": "List all posts",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "Page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Page size",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/post.PaginatedAllPosts"
                        }
                    },
                    "400": {
                        "description": "Page or limit is not a valid integer, or error fetching posts",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/post/user/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a paginated list of the posts authored by the user in the path id, filtered by the anonymous flag. The page query parameter is read as Page (capital P) and defaults to 1; limit defaults to 10 and is clamped to the range 1-100.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts"
                ],
                "summary": "List the posts of one user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Author user id (Mongo ObjectID hex)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "default": false,
                        "description": "Return anonymous posts",
                        "name": "anonymous",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "Page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Page size",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/post.PaginatedAllPosts"
                        }
                    },
                    "400": {
                        "description": "anonymous is not a boolean, Page or limit is not a valid integer, or error fetching posts",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/post/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the report post identified by the path id.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts"
                ],
                "summary": "Get a single post",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Post id (Mongo ObjectID hex)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/post.Post"
                        }
                    },
                    "400": {
                        "description": "invalid post id or post not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a report (lost and found, harassment, ragging, canteen overcharging, and so on) authored by the user in the path id. The payload is bound as a form: title (max 100 characters), description (max 5000 characters) and postType are required, and an optional image is validated, resized, compressed and uploaded to storage.",
                "consumes": [
                    "multipart/form-data",
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts"
                ],
                "summary": "Create a report post",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Author user id (Mongo ObjectID hex)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Post title (max 100 characters)",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Post description (max 5000 characters)",
                        "name": "description",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Report category",
                        "name": "postType",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Publish the report anonymously",
                        "name": "isAnonymous",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Optional image attachment",
                        "name": "image",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message and the created post.Post",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "invalid form body, validation failure, invalid image, upload failure or invalid user id",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes the report post identified by the path id. The caller role is taken from the access token and the owning user id is taken from the request body. Registered both as a user route and as an admin route.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts"
                ],
                "summary": "Delete a post",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Post id (Mongo ObjectID hex)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Owning user id",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/post.DeletePostRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message and the deleted post.Post",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "role missing from the context, invalid JSON body, or deletion failed",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates the report post identified by the path id. The payload is bound as a form, so send it as multipart/form-data or application/x-www-form-urlencoded. postType is required, and an optional replacement image is validated, resized, compressed and uploaded to storage.",
                "consumes": [
                    "multipart/form-data",
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts"
                ],
                "summary": "Update a post",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Post id (Mongo ObjectID hex)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Post title (max 100 characters)",
                        "name": "title",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Post description (max 5000 characters)",
                        "name": "description",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Report category",
                        "name": "postType",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Replacement image attachment",
                        "name": "image",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/post.Post"
                        }
                    },
                    "400": {
                        "description": "invalid form body, validation failure, invalid image, upload failure or update failed",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "post.DeletePostRequest": {
            "type": "object",
            "required": [
                "userID"
            ],
            "properties": {
                "userID": {
                    "type": "string"
                }
            }
        },
        "post.PaginatedAllPosts": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "posts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/post.Post"
                    }
                },
                "totalPages": {
                    "type": "integer"
                },
                "totalPosts": {
                    "type": "integer"
                }
            }
        },
        "post.Post": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "editHistory": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/post.UpdatePost"
                    }
                },
                "id": {
                    "type": "string"
                },
                "imageUrl": {
                    "type": "string"
                },
                "isAnonymous": {
                    "type": "boolean"
                },
                "postType": {
                    "type": "string"
                },
                "postedBy": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "post.UpdatePost": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "imageUrl": {
                    "type": "string"
                },
                "postType": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and the JWT access token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}